feat: modal with some logic
