+++ src/app/core/core.module.ts
@@ -5,13 +5,11 @@ import { HttpClient } from '@angular/common/http';
 import { HeaderComponent } from './header/header.component';
 import { httpLoaderFactory } from '../utils/http-loader.factory';
 import { AngularSvgIconModule } from 'angular-svg-icon';
-import { ModalComponent } from './modal/modal.component';
 
 
 @NgModule({
   declarations: [
     HeaderComponent,
-    ModalComponent
   ],
   imports: [
     CommonModule,
@@ -26,7 +24,6 @@ import { ModalComponent } from './modal/modal.component';
   ],
   exports: [
     HeaderComponent,
-    ModalComponent
   ]
 })
 export class CoreModule { }
