--- src/app/core/modal/modal.component.html
@@ -1,1 +1,0 @@
-Modal
