--- src/app/core/modal/modal.component.ts
@@ -1,15 +1,0 @@
-import { Component, ContentChild, OnInit, TemplateRef } from '@angular/core';
-
-@Component({
-  selector: 'app-grid',
-  templateUrl: './modal.component.html',
-  styleUrls: ['./modal.component.scss'],
-})
-export class ModalComponent implements OnInit {
-
-  constructor(
-  ) { }
-
-  ngOnInit(): void {}
-
-}
